# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Express ERP API",
    "description": "API documentation for Express ERP",
    "version": "1.0.0"
  },
  "host": "localhost:10000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Token must be included with 'Bearer' prefix.\n\nFormat: Bearer your-jwt-token\nExample: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n\nNote: The 'Bearer ' prefix is mandatory."
    }
  },
  "paths": {
...
    "/stockins": {
      "get": {
        "tags": [
          "Stockins"
        ],
        "summary": "Returns the list of all stockins",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "integer",
                    "description": "Stockin ID"
                  },
                  "stockinDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Date of stock in"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Additional notes"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          }
        }
      },
      "post": {
        "tags": [
          "Stockins"
        ],
        "summary": "Create a new stockin",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Stockin object that needs to be added",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "stockinDate": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Date of stock in"
                },
                "notes": {
                  "type": "string",
                  "description": "Additional notes"
                }
              },
              "required": [
                "stockinDate"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Stockin created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "ID": {
                  "type": "integer"
                },
                "stockinDate": {
                  "type": "string",
                  "format": "date-time"
                },
                "notes": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          }
        }
      }
    },
    "/stockins/{id}": {
      "get": {
        "tags": [
          "Stockins"
        ],
        "summary": "Get stockin by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of stockin to fetch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "ID": {
                  "type": "integer",
                  "description": "Stockin ID"
                },
                "stockinDate": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Date of stock in"
                },
                "notes": {
                  "type": "string",
                  "description": "Additional notes"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Stockin not found"
          }
        }
      },
      "put": {
        "tags": [
          "Stockins"
        ],
        "summary": "Update stockin by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of stockin to update"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Stockin object that needs to be updated",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "stockinDate": {
                  "type": "string",
                  "format": "date-time",
                  "description": "Date of stock in"
                },
                "notes": {
                  "type": "string",
                  "description": "Additional notes"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stockin updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "ID": {
                  "type": "integer"
                },
                "stockinDate": {
                  "type": "string",
                  "format": "date-time"
                },
                "notes": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Stockin not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Stockins"
        ],
        "summary": "Delete stockin by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of stockin to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Stockin deleted successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Stockin not found"
          }
        }
      }
    },
    "/detail-stockins": {
      "get": {
        "tags": [
          "Detail Stockins"
        ],
        "summary": "Returns the list of all detail stockins",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "ID": {
                    "type": "integer",
                    "description": "Detail stockin ID"
                  },
                  "StockinID": {
                    "type": "integer",
                    "description": "Associated stockin ID"
                  },
                  "productID": {
                    "type": "integer",
                    "description": "Product ID"
                  },
                  "quantity": {
                    "type": "integer",
                    "description": "Quantity of products"
                  },
                  "unitPrice": {
                    "type": "number",
                    "format": "float",
                    "description": "Price per unit"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          }
        }
      },
      "post": {
        "tags": [
          "Detail Stockins"
        ],
        "summary": "Create a new detail stockin",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Detail stockin object that needs to be added",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "StockinID": {
                  "type": "integer",
                  "description": "Associated stockin ID"
                },
                "productID": {
                  "type": "integer",
                  "description": "Product ID"
                },
                "quantity": {
                  "type": "integer",
                  "description": "Quantity of products"
                },
                "unitPrice": {
                  "type": "number",
                  "format": "float",
                  "description": "Price per unit"
                }
              },
              "required": [
                "StockinID",
                "productID",
                "quantity",
                "unitPrice"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Detail stockin created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "ID": {
                  "type": "integer"
                },
                "StockinID": {
                  "type": "integer"
                },
                "productID": {
                  "type": "integer"
                },
                "quantity": {
                  "type": "integer"
                },
                "unitPrice": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          }
        }
      }
    },
    "/detail-stockins/{id}": {
      "get": {
        "tags": [
          "Detail Stockins"
        ],
        "summary": "Get detail stockin by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of detail stockin to fetch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "ID": {
                  "type": "integer",
                  "description": "Detail stockin ID"
                },
                "StockinID": {
                  "type": "integer",
                  "description": "Associated stockin ID"
                },
                "productID": {
                  "type": "integer",
                  "description": "Product ID"
                },
                "quantity": {
                  "type": "integer",
                  "description": "Quantity of products"
                },
                "unitPrice": {
                  "type": "number",
                  "format": "float",
                  "description": "Price per unit"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Detail stockin not found"
          }
        }
      },
      "put": {
        "tags": [
          "Detail Stockins"
        ],
        "summary": "Update detail stockin by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of detail stockin to update"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Detail stockin object that needs to be updated",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "StockinID": {
                  "type": "integer",
                  "description": "Associated stockin ID"
                },
                "productID": {
                  "type": "integer",
                  "description": "Product ID"
                },
                "quantity": {
                  "type": "integer",
                  "description": "Quantity of products"
                },
                "unitPrice": {
                  "type": "number",
                  "format": "float",
                  "description": "Price per unit"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail stockin updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "ID": {
                  "type": "integer"
                },
                "StockinID": {
                  "type": "integer"
                },
                "productID": {
                  "type": "integer"
                },
                "quantity": {
                  "type": "integer"
                },
                "unitPrice": {
                  "type": "number",
                  "format": "float"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Detail stockin not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Detail Stockins"
        ],
        "summary": "Delete detail stockin by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of detail stockin to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Detail stockin deleted successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized - No token provided"
          },
          "403": {
            "description": "Forbidden - Invalid token"
          },
          "404": {
            "description": "Detail stockin not found"
          }
        }
      }
    },
...
  }
}